Organised tests under correct folders
