+++ src/Syringe.Tests/Syringe.Tests.csproj
@@ -181,12 +181,12 @@
     <Compile Include="StubsMocks\TestCaseSessionRepositoryMock.cs" />
     <Compile Include="Integration\Git\GitProviderTests.cs" />
     <Compile Include="Unit\Core\Tests\VariableTests.cs" />
-    <Compile Include="Unit\Git\GitProviderTests.cs" />
-    <Compile Include="Unit\IO\FIleHandlerTests.cs" />
-    <Compile Include="Unit\Helpers\UrlHelperTests.cs" />
+    <Compile Include="Unit\Core\Git\GitProviderTests.cs" />
+    <Compile Include="Unit\Core\IO\FIleHandlerTests.cs" />
+    <Compile Include="Unit\Core\Helpers\UrlHelperTests.cs" />
     <Compile Include="Unit\Web\MvcConfigurationTests.cs" />
     <Compile Include="Unit\Web\TestFileMapperTests.cs" />
-    <Compile Include="Unit\Repositories\TestRepositoryTests.cs" />
+    <Compile Include="Unit\Core\Repositories\TestRepositoryTests.cs" />
     <Compile Include="Unit\Web\ExtensionsTests.cs" />
     <Compile Include="Unit\Web\HomeControllerTests.cs" />
     <Compile Include="Unit\Web\JsonControllerTests.cs" />
@@ -194,25 +194,25 @@
     <Compile Include="Unit\Web\RunViewModelTests.cs" />
     <Compile Include="Unit\Web\TestControllerTests.cs" />
     <Compile Include="Unit\Web\TestFileControllerTests.cs" />
-    <Compile Include="Unit\Repositories\Xml\Writer\TestFileWriterTests.cs" />
+    <Compile Include="Unit\Core\Repositories\Xml\Writer\TestFileWriterTests.cs" />
     <Compile Include="Integration\Xml\TestFileRunnerTests.cs" />
     <Compile Include="TestHelpers.cs" />
-    <Compile Include="Unit\Http\HttpResponseTests.cs" />
+    <Compile Include="Unit\Core\Http\HttpResponseTests.cs" />
     <Compile Include="StubsMocks\RestResponseStub.cs" />
     <Compile Include="StubsMocks\RestClientMock.cs" />
-    <Compile Include="Unit\Http\HttpClientTests.cs" />
-    <Compile Include="Unit\Results\TestResultTests.cs" />
-    <Compile Include="Unit\Runner\CapturedVariableProviderTests.cs" />
-    <Compile Include="Unit\Runner\SessionVariablesTests.cs" />
-    <Compile Include="Unit\Runner\TestFileResultsBuilder.cs" />
-    <Compile Include="Unit\Runner\TestFileResultTests.cs" />
-    <Compile Include="Unit\Runner\VerificationsMatcherTests.cs" />
+    <Compile Include="Unit\Core\Http\HttpClientTests.cs" />
+    <Compile Include="Unit\Core\Tests\Results\TestResultTests.cs" />
+    <Compile Include="Unit\Core\Runner\CapturedVariableProviderTests.cs" />
+    <Compile Include="Unit\Core\Runner\SessionVariablesTests.cs" />
+    <Compile Include="Unit\Core\Runner\TestFileResultsBuilder.cs" />
+    <Compile Include="Unit\Core\Runner\TestFileResultTests.cs" />
+    <Compile Include="Unit\Core\Runner\VerificationsMatcherTests.cs" />
     <Compile Include="StubsMocks\TestCaseReaderMock.cs" />
-    <Compile Include="Unit\Runner\TestSessionRunnerTests.cs" />
-    <Compile Include="Unit\Http\HttpLogWriterTests.cs" />
+    <Compile Include="Unit\Core\Runner\TestSessionRunnerTests.cs" />
+    <Compile Include="Unit\Core\Http\HttpLogWriterTests.cs" />
     <Compile Include="StubsMocks\HttpClientMock.cs" />
-    <Compile Include="Unit\Repositories\Xml\Reader\TestFileReaderTests.cs" />
-    <Compile Include="Unit\Xml\XmlHelperTests.cs" />
+    <Compile Include="Unit\Core\Repositories\Xml\Reader\TestFileReaderTests.cs" />
+    <Compile Include="Unit\Core\Xml\XmlHelperTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -234,7 +234,7 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Unit\Repositories\Xml\Writer\XmlExamples\variables.xml">
+    <EmbeddedResource Include="Unit\Core\Repositories\Xml\Writer\XmlExamples\variables.xml">
       <SubType>Designer</SubType>
     </EmbeddedResource>
     <None Include="app.config" />
@@ -253,28 +253,28 @@
       <SubType>Designer</SubType>
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </EmbeddedResource>
-    <EmbeddedResource Include="Unit\Repositories\Xml\Reader\XmlExamples\full.xml" />
-    <EmbeddedResource Include="Unit\Repositories\Xml\Reader\XmlExamples\single-test.xml">
+    <EmbeddedResource Include="Unit\Core\Repositories\Xml\Reader\XmlExamples\full.xml" />
+    <EmbeddedResource Include="Unit\Core\Repositories\Xml\Reader\XmlExamples\single-test.xml">
       <SubType>Designer</SubType>
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Unit\Repositories\Xml\Writer\XmlExamples\all-attributes.xml">
+    <EmbeddedResource Include="Unit\Core\Repositories\Xml\Writer\XmlExamples\all-attributes.xml">
       <SubType>Designer</SubType>
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Unit\Repositories\Xml\Writer\XmlExamples\headers.xml">
+    <EmbeddedResource Include="Unit\Core\Repositories\Xml\Writer\XmlExamples\headers.xml">
       <SubType>Designer</SubType>
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Unit\Repositories\Xml\Writer\XmlExamples\postbody.xml">
+    <EmbeddedResource Include="Unit\Core\Repositories\Xml\Writer\XmlExamples\postbody.xml">
       <SubType>Designer</SubType>
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Unit\Repositories\Xml\Writer\XmlExamples\capturedvariables.xml">
+    <EmbeddedResource Include="Unit\Core\Repositories\Xml\Writer\XmlExamples\capturedvariables.xml">
       <SubType>Designer</SubType>
     </EmbeddedResource>
   </ItemGroup>
@@ -282,13 +282,13 @@
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Unit\Repositories\Xml\Writer\XmlExamples\assertions.xml">
+    <EmbeddedResource Include="Unit\Core\Repositories\Xml\Writer\XmlExamples\assertions.xml">
       <SubType>Designer</SubType>
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
     <Content Include="Integration\Xml\XmlExamples\Runner\50-cases.xml" />
-    <EmbeddedResource Include="Unit\Repositories\Xml\Writer\XmlExamples\large-file.xml">
+    <EmbeddedResource Include="Unit\Core\Repositories\Xml\Writer\XmlExamples\large-file.xml">
       <SubType>Designer</SubType>
     </EmbeddedResource>
   </ItemGroup>
