+++ src/Syringe.Tests/Unit/Core/Repositories/Xml/Reader/XmlExamples/full.xml
