+++ src/Syringe.Tests/Unit/Core/Repositories/Xml/Reader/XmlExamples/single-test.xml
