+++ src/Syringe.Tests/Unit/Core/Repositories/Xml/Writer/XmlExamples/all-attributes.xml
