+++ src/Syringe.Tests/Unit/Core/Repositories/Xml/Writer/XmlExamples/assertions.xml
