+++ src/Syringe.Tests/Unit/Core/Repositories/Xml/Writer/XmlExamples/large-file.xml
