+++ src/Syringe.Tests/Unit/Core/Repositories/Xml/Writer/XmlExamples/postbody.xml
