+++ src/Syringe.Tests/Unit/Core/Xml/XmlHelperTests.cs
@@ -1,11 +1,10 @@
 ﻿using System.Linq;
 using System.Xml.Linq;
 using NUnit.Framework;
-using Syringe.Core.Repositories.XML.Reader;
 using Syringe.Core.Xml;
 using XmlException = Syringe.Core.Exceptions.XmlException;
 
-namespace Syringe.Tests.Unit.Xml
+namespace Syringe.Tests.Unit.Core.Xml
 {
     public class XmlHelperTests
     {
